+++ structs_lwh/mbr.go
@@ -5,16 +5,16 @@ type Partitions struct {
 	PartStatus byte
 	PartType   byte
 	PartFit    byte
-	PartStart  int32
-	PartSize   int32
+	PartStart  int64
+	PartSize   int64
 	PartName   [16]byte
 }
 
 // MBR estructura para el master boot record
 type MBR struct {
-	MbrSize          int32
+	MbrSize          int64
 	MbrTime          [25]byte
-	MbrDiskSignature int32
+	MbrDiskSignature int64
 	Partition        [4]Partitions
 }
